+++ packages/collapse/CHANGELOG.md
@@ -3,9 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-# 0.1.0 (2022-08-03)
+# [1.1.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/collapse@0.1.0...@heycar-uikit/collapse@1.1.0) (2022-09-01)
 
 
 ### Features
 
-* **collapse:** the first version of the collapse component ([#46](https://github.com/hey-car/heycar-uikit/issues/46)) ([4d10244](https://github.com/hey-car/heycar-uikit/commit/4d102446b2c6348408bb3bd1d5a7ebf1ec1cd6f5))
+* **collapse:** the first version of the collapse component ([#51](https://github.com/hey-car/heycar-uikit/issues/51)) ([b8d3f3f](https://github.com/hey-car/heycar-uikit/commit/b8d3f3f88cdfde98bb0f6364973895a5e9969182))
+
+
+
+
+
+# 0.1.0 (2022-08-03)
+
+### Features
+
+- **collapse:** the first version of the collapse component ([#46](https://github.com/hey-car/heycar-uikit/issues/46)) ([4d10244](https://github.com/hey-car/heycar-uikit/commit/4d102446b2c6348408bb3bd1d5a7ebf1ec1cd6f5))