+++ packages/fonts/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/fonts@1.1.0...@heycar-uikit/fonts@1.3.0) (2022-09-01)
+
+
+### Features
+
+* **badge:** The first implantation of badge component ([#42](https://github.com/hey-car/heycar-uikit/issues/42)) ([2ca5a7e](https://github.com/hey-car/heycar-uikit/commit/2ca5a7ea897c4cdeccaadca4f4d1592be40fe6e7))
+* **collapse:** the first version of the collapse component ([#51](https://github.com/hey-car/heycar-uikit/issues/51)) ([b8d3f3f](https://github.com/hey-car/heycar-uikit/commit/b8d3f3f88cdfde98bb0f6364973895a5e9969182))
+* **icons:** the first implementation of the icons package ([#39](https://github.com/hey-car/heycar-uikit/issues/39)) ([74fd00d](https://github.com/hey-car/heycar-uikit/commit/74fd00d9227e3e915fbb54962d56ab9d64715b5c))
+* **switch:** initial implementation of the switch component ([#41](https://github.com/hey-car/heycar-uikit/issues/41)) ([fae4001](https://github.com/hey-car/heycar-uikit/commit/fae40014e387fd6a84275fdee825b265b798b7d3))
+
+
+
+
+
 # [1.1.0](https://github.com/hey-car/heycar-uikit/compare/@heycar-uikit/fonts@1.1.0...@heycar-uikit/fonts@1.1.0) (2022-08-03)